requirements.txt
tensorflow-cpu==2.12.0
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.2
tensorboard==2.12.1
matplotlib==3.7.1
streamlit==1.22.0
scikeras==0.11.0
nltk==3.8.1
